refactor(mcp): save intermediate progress before pure architecture refactoring
